Implement Scoreboard system and utilities
This update includes implementation of a lightweight packet-based scoreboard system, including a way to register the tasks to the loader. It also includes the creation of utility classes to handle scoreboard based tasks. The newly implemented scoreboard system can be used asynchronously. It also provides better abstraction and code management.